MS/requirements.txt
pandas
numpy
scipy
psycopg2-binary
django
sqlalchemy
djangorestframework
requests
django-cors-headers
sklearn
pyspark
PyArrow
